predict_service/requirements.txt
Django==4.1.5
djangorestframework==3.14.0
psycopg2-binary==2.8.6
python-dotenv==0.21.1
cryptography==39.0.1
numpy